Fix TCP keepalive idle_time_sec, interval_sec type #432

+++ clickhouse_driver/client.py
@@ -872,7 +872,7 @@ class Client(object):
                 except ValueError:
                     parts = value.split(',')
                     kwargs[name] = (
-                        float(parts[0]), float(parts[1]), int(parts[2])
+                        int(parts[0]), int(parts[1]), int(parts[2])
                     )
             elif name == 'client_revision':
                 kwargs[name] = int(value)
